requirements.txt
awscli>=1.22.11
feedgenerator>=2.0.0
feedparser>=6.0.8
python-decouple>=3.5
pytest>=6.2.5
elevate>=0.1.3
requests>=2.31.0
python-dotenv>=0.19.2
rich>=12.6.0
boto3>=1.26.151
# Linux only
hunspell